+++ wwt_data_formats/__init__.py
@@ -159,14 +159,9 @@ def _parse_trait(trait_spec, text):
         raise ValueError(f'cannot interpret text {text!r} as a boolean')
 
     if isinstance(trait_spec, UseEnum):
-        # We could do better than a linear search here, but meh.
         if text is None:
             text = ''
-
-        for enum_value in trait_spec.enum_class:
-            if text == enum_value.value:
-                return enum_value
-        raise ValueError(f'unrecognized value {text!r} for enumeration {trait_spec.enum_class}')
+        return trait_spec.enum_class.from_text(text)
 
     raise ValueError(f'internal error: unimplemented parse for trait type {trait_spec}')
 
@@ -216,7 +211,8 @@ class LockedXmlTraits(LockedDownTraits):
 
             if xml_spec == XmlSer.ATTRIBUTE:
                 text = elem.attrib.get(xml_data[0])
-                value = _parse_trait(tspec, text)
+                if text is not None:
+                    value = _parse_trait(tspec, text)
             elif xml_spec == XmlSer.TEXT_ELEM:
                 sub = elem.find(xml_data[0])
                 if sub is not None:
@@ -435,6 +431,51 @@ class LockedXmlTraits(LockedDownTraits):
 
         """
         return self._serialize_xml(None)
+
+
+    def write_xml(self, dest_stream, dest_wants_bytes=False, indent=True):
+        """Serialize this object to XML, writing the data to a stream.
+
+        Parameters
+        ----------
+        dest_stream : writeable file-like object or None
+          The destination to which the XML data will be written. If None,
+          standard output is used.
+        dest_wants_bytes : optional bool, default False
+          Whether the destination stream expects to be fed bytes data rather
+          than Unicode. If so, the XML text is encoded into UTF-8 before
+          being written.
+        indent : bool, default True
+          Whether the returned XML text will have user-friendly indentation or not.
+
+        Returns
+        -------
+        None
+
+        """
+        write_xml_doc(
+            self.to_xml(),
+            dest_stream = dest_stream,
+            dest_wants_bytes = dest_wants_bytes,
+            indent = indent,
+        )
+
+
+    def to_xml_string(self, indent=True):
+        """Serialize this object to XML text.
+
+        Parameters
+        ----------
+        indent : optional bool, default True
+          Whether the returned XML text will have user-friendly indentation or not.
+
+        Returns
+        -------
+        xml_text : string
+          A textual serialization of the object as XML.
+
+        """
+        return stringify_xml_doc(self.to_xml(), indent=indent)
 
 
 def indent_xml(elem, level=0):