+++ CHANGES.md
@@ -1,4 +1,4 @@
-# Next version (unreleased)
+# 0.1.2 (2020-Mar-20)
 
 - Fix roundtripping of imagesets with a non-empty QuadTreeMap, and add a few
   more attributes on the `<ImageSet>` type.